refactor(codegen): extract compile_expr into expressions.rs (#113 phase 6)
